requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
fpdf
folium
pytesseract
matplotlib